+++ exonum-node/src/consensus.rs
@@ -14,7 +14,7 @@
 
 use anyhow::bail;
 use exonum::{
-    blockchain::{contains_transaction, Blockchain, BlockchainMut, ProposerId, Schema},
+    blockchain::{Blockchain, BlockchainMut, PersistentPool, ProposerId, Schema, TransactionCache},
     crypto::{Hash, PublicKey},
     helpers::{Height, Round, ValidatorId},
     merkledb::{BinaryValue, Fork, ObjectHash, Patch},
@@ -292,15 +292,11 @@ impl NodeHandler {
             let block_height = block.height;
             let incomplete_block = IncompleteBlock::new(block, transactions, precommits);
             let snapshot = self.blockchain.snapshot();
-            let schema = Schema::new(&snapshot);
+            let txs_pool = Schema::new(snapshot.as_ref()).transactions_pool();
 
             let has_unknown_txs = self
                 .state
-                .create_incomplete_block(
-                    incomplete_block,
-                    &schema.transactions(),
-                    &schema.transactions_pool(),
-                )
+                .create_incomplete_block(incomplete_block, &snapshot, &txs_pool)
                 .has_unknown_txs();
 
             let known_nodes = self.remove_request(&RequestData::Block(block_height));
@@ -675,40 +671,35 @@ impl NodeHandler {
     ) {
         trace!("COMMIT {:?}", block_hash);
 
-        // Merge changes into storage
-        let (committed_txs, proposer) = {
-            let (committed_txs, proposer) = {
-                let block_state = self.state.block_mut(&block_hash).unwrap();
-                let committed_txs = block_state.txs().len();
-                let proposer = block_state.proposer_id();
+        let mut block_state = self.state.take_block_for_commit(&block_hash);
+        let committed_txs = block_state.txs();
+        let proposer = block_state.proposer_id();
 
-                self.blockchain
-                    .commit(
-                        block_state.patch(),
-                        block_hash,
-                        precommits,
-                        self.state.tx_cache_mut(),
-                    )
-                    .expect("Cannot commit block");
+        // Remove committed transactions from the cache.
+        for tx_hash in committed_txs {
+            self.state.tx_cache_mut().remove(tx_hash);
+        }
+        let committed_txs_len = committed_txs.len();
 
-                // Consensus messages cache is useful only during one height, so it should be
-                // cleared when a new height is achieved.
-                self.blockchain.persist_changes(
-                    |schema| schema.consensus_messages_cache().clear(),
-                    "Cannot clear consensus messages",
-                );
+        // Commit block patch to the storage.
+        self.blockchain
+            .commit(block_state.patch(), block_hash, precommits)
+            .expect("Cannot commit block");
 
-                (committed_txs, proposer)
-            };
-            // Update node state.
-            self.state
-                .update_config(Schema::new(&self.blockchain.snapshot()).consensus_config());
-            // Update state to new height.
-            let block_hash = self.blockchain.as_ref().last_hash();
-            self.state
-                .new_height(&block_hash, self.system_state.current_time());
-            (committed_txs, proposer)
-        };
+        // Consensus messages cache is useful only during one height, so it should be
+        // cleared when a new height is achieved.
+        self.blockchain.persist_changes(
+            |schema| schema.consensus_messages_cache().clear(),
+            "Cannot clear consensus messages",
+        );
+
+        // Update node state.
+        self.state
+            .update_config(Schema::new(&self.blockchain.snapshot()).consensus_config());
+        // Update state to new height.
+        let block_hash = self.blockchain.as_ref().last_hash();
+        self.state
+            .new_height(&block_hash, self.system_state.current_time());
 
         let snapshot = self.blockchain.snapshot();
         for plugin in &self.plugins {
@@ -724,7 +715,7 @@ impl NodeHandler {
             height,
             proposer,
             round.map_or_else(|| "?".to_owned(), |x| x.to_string()),
-            committed_txs,
+            committed_txs_len,
             pool_len,
             block_hash.to_hex(),
         );
@@ -761,8 +752,8 @@ impl NodeHandler {
         let hash = msg.object_hash();
 
         let snapshot = self.blockchain.snapshot();
-        let schema = Schema::new(&snapshot);
-        if contains_transaction(&hash, &schema.transactions(), self.state.tx_cache()) {
+        let tx_pool = PersistentPool::new(&snapshot, self.state.tx_cache());
+        if tx_pool.contains_transaction(hash) {
             return Err(HandleTxError::AlreadyProcessed);
         }
 
@@ -780,8 +771,16 @@ impl NodeHandler {
             // Thus, we don't stop the execution here.
             outcome = Err(HandleTxError::Invalid(e));
         } else {
-            // Transaction is OK, store it to the cache.
-            self.state.tx_cache_mut().insert(hash, msg);
+            // Transaction is OK, store it to the cache or persistent pool.
+            if self.state.persist_txs_immediately() {
+                let fork = self.blockchain.fork();
+                Schema::new(&fork).add_transaction_into_pool(msg);
+                self.blockchain
+                    .merge(fork.into_patch())
+                    .expect("Cannot add transaction to persistent pool");
+            } else {
+                self.state.tx_cache_mut().insert(hash, msg);
+            }
             outcome = Ok(());
         }
 
@@ -1061,12 +1060,8 @@ impl NodeHandler {
         height: Height,
         tx_hashes: &[Hash],
     ) -> (Hash, Patch) {
-        self.blockchain.create_patch(
-            proposer_id,
-            height,
-            tx_hashes,
-            &mut self.state.tx_cache_mut(),
-        )
+        self.blockchain
+            .create_patch(proposer_id, height, tx_hashes, self.state.tx_cache())
     }
 
     /// Calls `create_block` with transactions from the corresponding `Propose` and returns the